+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * feat(TypeB24.callMethod/TypeHttp.call): deprecate param `start`, now use param `params.start`
 * feat(TypeB24.callMethod/TypeB24.callFastListMethod/TypeB24.fetchListMethod/TypeB24.callBatch/TypeB24.callBatchByChunk)!: add param requestId
 * feat(RestrictionManager): new restrictions
+* feat(PlacementManager): use the property name `placement` instead of `title`
 
 ## [0.5.1](https://github.com/bitrix24/b24jssdk/compare/v0.4.10...v0.5.1) (2025-10-29)